+++ src/main/java/com/timlis/pojo/Blog.java
@@ -18,7 +18,7 @@ public class Blog {
     private String content;
     private String firstPicture;
     private String flag;
-    private Integer views;
+    private Double views;
     private boolean appreciation;
     private boolean shareStatement;
     private boolean commentabled;
@@ -123,11 +123,11 @@ public class Blog {
         this.flag = flag;
     }
 
-    public Integer getViews() {
+    public Double getViews() {
         return views;
     }
 
-    public void setViews(Integer views) {
+    public void setViews(Double views) {
         this.views = views;
     }
